% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\docType{data}
\name{matches}
\alias{matches}
\title{Matches}
\format{
A data frame with 21 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{season_id}{\code{text}. The unique ID number for the season.
References \code{season_id} in the \code{seasons} dataset.}
\item{season}{\code{integer}. The year that the season started.}
\item{tier}{\code{integer}. The tier in English football. The possible
values are \code{1} through \code{4}.}
\item{division}{\code{test}. The division in English football. For the
current league structure, the possible values are \code{Premier League},
\code{Championship}, \code{League One}, and \code{League Two}. For previous
league structures, the possible values are \code{First Division},
\code{Second Division}, \code{Third Division}, and \code{Fourth Division}.}
\item{subdivision}{\code{test}. The subdivision in English football. In the
third tier, from the 1921-22 season through the 1957-58 season, there were
North and South subdivisions. The possible values are \code{North},
\code{South}, and \code{None}.}
\item{match_id}{\code{text}. The unique ID number for the match. Has the
format \code{M-####-###}, where the first number is the season and second
number is a within-season counter that is assigned with the data sorted by
the name of the home team, then by the name of the away team.}
\item{match_name}{\code{text}. The name of the match.}
\item{home_team_id}{\code{text}. The unique ID number for the home team.
References \code{team_id} in the \code{teams} dataset.}
\item{home_team_name}{\code{text}. The name of the home team. See the
\code{teams} dataset.}
\item{away_team_id}{\code{text}. The unique ID number for the away team.
References \code{team_id} in the \code{teams} dataset.}
\item{away_team_name}{\code{text}. The name of the away team. See the
\code{teams} dataset.}
\item{score}{\code{text}. The score of the match in the format \code{#-#},
where the first number is the score of the home team and the second number
is the score of the away team.}
\item{home_team_score}{\code{integer}. The score of the home team.}
\item{away_team_score}{\code{integer}. The score of the away team.}
\item{home_team_score_margin}{\code{integer}. The score margin for the home
team.}
\item{away_team_score_margin}{\code{integer}. The score margin for the away
team.}
\item{result}{\code{enum}. The result of the match. The possible values are
\code{home team win}, \code{away team win}, and \code{draw}.}
\item{home_team_win}{\code{boolean}. Whether the home team won the match.
Coded \code{1} if the home team won the match and \code{0} otherwise.}
\item{away_team_win}{\code{boolean}. Whether the home team won the match.
Coded \code{1} if the home team won the match and \code{0} otherwise.}
\item{draw}{\code{boolean}. Whether the match ended in a draw. Coded
\code{1} of the match ended in a draw and \code{0} otherwise.}
}
}
\usage{
matches
}
\description{
This dataset records all matches that have ever been played in the English
Football League and the Premier League (1888-2022). There is one
observation per match per season. It indicates the season, tier, division,
and subdivision for the match, the score, the score margin for each team,
and the result of the match (home team win, away team win, draw).
}
\keyword{datasets}
