% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\docType{data}
\name{standings}
\alias{standings}
\title{Standings}
\format{
A data frame with 18 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{season_id}{\code{text}. The unique ID number for the season.
References \code{season_id} in the \code{seasons} dataset.}
\item{season}{\code{integer}. The year that the season started.}
\item{tier}{\code{integer}. The tier in English football. The possible
values are \code{1} through \code{4}.}
\item{division}{\code{test}. The division in English football. For the
current league structure, the possible values are \code{Premier League},
\code{Championship}, \code{League One}, and \code{League Two}. For previous
league structures, the possible values are \code{First Division},
\code{Second Division}, \code{Third Division}, and \code{Fourth Division}.}
\item{subdivision}{\code{test}. The subdivision in English football. In the
third tier, from the 1921-22 season through the 1957-58 season, there were
North and South subdivisions. The possible values are \code{North},
\code{South}, and \code{None}.}
\item{position}{\code{integer}. The team's final position in the league.}
\item{team_id}{\code{text}. The unique ID number for the team. References
\code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team.}
\item{played}{\code{integer}. The number of matches that the team played.}
\item{wins}{\code{integer}. The number of matches that the team won.}
\item{draws}{\code{integer}. The number of matches that the team drew.}
\item{losses}{\code{integer}. The number of matches that the team lost.}
\item{goals_for}{\code{integer}. The number of goals scored by the team.}
\item{goals_against}{\code{integer}. The number of goals scored against the
team.}
\item{goal_difference}{\code{integer}. The team's goal difference.}
\item{points}{\code{integer}. The number of points that the team earned
over the whole season (after any point adjustments).}
\item{point_adjustment}{\code{integer}. The number of points that were
deducted by the league due to violations of league rules or added by the
league due to forfeits.}
}
}
\usage{
standings
}
\description{
This dataset records all end-of-the-season standings. There is one
observation per team per season. It indicates the final position of the
team (accounting for tie-breakers), the name of the team, the number of
matches played, the number of wins, the number of losses, the number of
draws, the number of goals for, the number of goals against, the goal
difference, and the total number of points earned.
}
\keyword{datasets}
